Refine generated route and action typings
