+++ resources/js/features/manage/admin/dashboard/types.ts
@@ -1,0 +1,5 @@
+export type TranslatorFn = (
+    key: string,
+    fallback?: string | null,
+    replacements?: Record<string, unknown>
+) => string;
