+++ resources/js/types/generated.d.ts
@@ -2,7 +2,17 @@
 // 於測試環境中並不會實際存在，因此在此以 any 型別提供最小化的型別定義。
 
 declare module '@/actions/*' {
-    const controller: (...args: unknown[]) => unknown;
+    export type ActionRouteMethod = 'get' | 'post' | 'put' | 'patch' | 'delete' | 'GET' | 'POST' | 'PUT' | 'PATCH' | 'DELETE';
+    export type ActionRouteDefinition = { url: string; method: ActionRouteMethod };
+    export interface ActionRouteInvoker {
+        (...args: unknown[]): ActionRouteDefinition;
+        url: string;
+        method: string;
+    }
+    export type ActionController = ((...args: unknown[]) => unknown) & {
+        [key: string]: ActionRouteInvoker;
+    };
+    const controller: ActionController;
     export default controller;
 }
 
@@ -14,22 +24,25 @@ declare module '@/routes/*' {
     export type RouteFormDefinition<Method extends string | readonly string[] = string> = Method extends readonly string[]
         ? { action: string; method: Method[number] }
         : { action: string; method: Method };
-    export type RouteInvoker = {
+    export interface RouteFunction {
         (...args: unknown[]): string;
         url: (options?: RouteQueryOptions) => string;
         form: (...args: unknown[]) => RouteFormDefinition;
+    }
+    export type RouteInvoker = RouteFunction & {
+        [key: string]: RouteInvoker;
     };
     export function url(options?: RouteQueryOptions): string;
-    export const form: (...args: unknown[]) => RouteFormDefinition;
-    export const get: (...args: unknown[]) => RouteDefinition;
-    export const post: (...args: unknown[]) => RouteDefinition;
-    export const put: (...args: unknown[]) => RouteDefinition;
-    export const patch: (...args: unknown[]) => RouteDefinition;
-    export const head: (...args: unknown[]) => RouteDefinition;
-    export const destroy: (...args: unknown[]) => RouteDefinition;
-    export const edit: (...args: unknown[]) => RouteDefinition;
-    export const request: (...args: unknown[]) => RouteDefinition;
-    export const send: (...args: unknown[]) => RouteDefinition;
+    export const form: RouteFunction['form'];
+    export const get: RouteInvoker;
+    export const post: RouteInvoker;
+    export const put: RouteInvoker;
+    export const patch: RouteInvoker;
+    export const head: RouteInvoker;
+    export const destroy: RouteInvoker;
+    export const edit: RouteInvoker;
+    export const request: RouteInvoker;
+    export const send: RouteInvoker;
     const route: RouteInvoker;
     export default route;
 }
